# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MiniMarket Plus API",
    "description": "Backend para gestion de inventario, ventas, carrito y usuarios con seguridad JWT.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "Autenticacion"
        ],
        "summary": "Registra un cliente con rol ROLE_CLIENTE",
        "description": "Crea una cuenta cliente, cifra la contrasena con BCrypt y retorna un token JWT para el nuevo usuario.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Cliente registrado correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "409": {
            "description": "El username ya existe"
          },
          "500": {
            "description": "Error interno del servidor"
          },
          "400": {
            "description": "Solicitud invalida"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Autenticacion"
        ],
        "summary": "Inicia sesion y retorna un token JWT",
        "description": "Autentica un usuario registrado mediante username y password. Si las credenciales son validas, retorna token JWT, tipo Bearer, username y roles asociados.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Inicio de sesion correcto",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Credenciales invalidas"
          },
          "500": {
            "description": "Error interno del servidor"
          },
          "400": {
            "description": "Solicitud invalida"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RegisterRequest": {
        "type": "object",
        "description": "Datos requeridos para registrar un nuevo cliente",
        "properties": {
          "username": {
            "type": "string",
            "description": "Nombre de usuario nuevo",
            "example": "cliente_nuevo",
            "maxLength": 50,
            "minLength": 3
          },
          "password": {
            "type": "string",
            "description": "Contrasena del nuevo usuario",
            "example": "Cliente123",
            "maxLength": 100,
            "minLength": 6
          }
        },
        "required": [
          "password",
          "username"
        ]
      },
      "AuthResponse": {
        "type": "object",
        "description": "Respuesta de autenticacion generada luego de login o registro exitoso",
        "properties": {
          "token": {
            "type": "string",
            "description": "Token JWT que debe enviarse como Bearer token en endpoints protegidos",
            "example": "eyJhbGciOiJIUzI1NiJ9..."
          },
          "tipo": {
            "type": "string",
            "description": "Tipo de token utilizado por el esquema de autenticacion",
            "example": "Bearer"
          },
          "username": {
            "type": "string",
            "description": "Nombre del usuario autenticado",
            "example": "admin"
          },
          "roles": {
            "type": "array",
            "description": "Roles asociados al usuario autenticado",
            "example": [
              "ROLE_ADMIN"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "description": "Credenciales requeridas para iniciar sesion y obtener un token JWT",
        "properties": {
          "username": {
            "type": "string",
            "description": "Nombre de usuario registrado",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "description": "Contrasena del usuario registrado",
            "example": "Admin123"
          }
        },
        "required": [
          "password",
          "username"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}